+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.9.0] - 2023-10-31
+## [0.9.0] - 2023-11-01
 
 ### Changed
 
@@ -19,6 +19,8 @@ channels can be added to the canvas through a gallery click
 assumption that the current mask should be applied to a different image
 - Cache the quantification measurements CSV and UMAP coordinates
 as Serverside objects to speed up interactive UMAP for large datasets
+- Annotations linked to quantification sheets are parsed to identify
+which column should be used to match the annotation to the ROI
 
 ### Added
 
@@ -32,6 +34,8 @@ can now switch on scroll zoom on the canvas (default is not enabled)
 cell counting
 - Link between subsetting cells on the UMAP and sending a query to
 the query tab to display relevant ROIs by descending cell count
+- Pattern matching on mask and ROI names to automatically find the
+proper mask on an ROI change
 
 ### Fixed